airflow-db/requirements.txt
duckdb
apache-airflow-providers-amazon
pytest
pytest-black
flake8
sqlfluff
click